clean up probability config and start merging tests
