+++ hydra/src/openset/CMakeLists.txt
@@ -4,5 +4,4 @@ target_sources(
           ${CMAKE_CURRENT_SOURCE_DIR}/embedding_group.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/embedding_pooling.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/clustering/agglomerative_clustering.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/clustering/ib_utilities.cpp
 )
